Improve consistency in Range/Rect/Box classes.
More misc refactoring.

core:
- Range/Rect/Box classes now have a consistent API and no longer hide
  their members.
- Removed unnecessary initializer_list constructor from Vector classes.
- Can now crop Array1DViews.

CUDA:
- Moved ArrayUtils into free functions inside a namespace.
- Added box-line intersection.
- Renamed DeviceVector to DeviceArray1D.
- Cleaned up constructors for DeviceArray*D.
- Moved MathUtils and ThreadMath functions into their own namespaces.

+++ CUDA/src/KernelArray3D.inl
@@ -16,7 +16,7 @@ KernelArray3D< T >::KernelArray3D( cudaPitchedPtr d_pitchedPointer, int depth ) 
 
     md_pPitchedPointer( reinterpret_cast< T* >( d_pitchedPointer.ptr ) ),
     m_width( static_cast< int >( d_pitchedPointer.xsize / sizeof( T ) ) ),
-    m_height( d_pitchedPointer.ysize ),
+    m_height( static_cast< int >( d_pitchedPointer.ysize ) ),
     m_depth( depth ),
     m_rowPitch( d_pitchedPointer.pitch )
